main backend done

+++ routes/warrantys.js
@@ -1,7 +1,9 @@
 const express = require('express');
+const { ObjectID } = require('mongodb');
 const router = express.Router();
 
 const catchAsync = require('../utils/catchAsync');
+const { dateToISOLikeButLocal } = require('../utils/date');
 const ExpressError = require('../utils/ExpressError');
 const Warranty = require('../models/warranty');
 const { warrantySchema } = require('../schemas.js');
@@ -11,7 +13,11 @@ router.get(
   '/',
   isLoggedIn,
   catchAsync(async (req, res) => {
-    const warrantys = await Warranty.find({});
+    const userId = req.user._doc._id;
+
+    const warrantys = await Warranty.find({
+      owner: userId,
+    });
     res.render('warrantys/index', { warrantys });
   })
 );
@@ -37,7 +43,10 @@ router.post(
   isLoggedIn,
   validateWarranty,
   catchAsync(async (req, res) => {
-    const warranty = new Warranty(req.body.warranty);
+    const warranty = new Warranty({
+      ...req.body.warranty,
+      owner: req.user._doc._id,
+    });
     await warranty.save();
     req.flash('success', 'Successfully maded a warranty card');
     res.redirect('/warrantys');
@@ -50,7 +59,7 @@ router.get(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const warranty = await Warranty.findById(req.params.id);
-    if (!warranty) {
+    if (!warranty || !warranty.owner.equals(req.user._doc._id)) {
       req.flash('error', 'Cannot find the warranty card');
       return res.redirect('/warrantys');
     }
@@ -64,8 +73,15 @@ router.get(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const warranty = await Warranty.findById(req.params.id);
+    if (!warranty.owner.equals(req.user._doc._id)) {
+      req.flash('error', 'Cannot find the warranty card');
+      return res.redirect('/warrantys');
+    }
 
-    res.render('warrantys/edit', { warranty });
+    res.render('warrantys/edit', {
+      warranty,
+      purchase: dateToISOLikeButLocal(warranty.purchase),
+    });
   })
 );
 
@@ -79,6 +95,10 @@ router.put(
     const warranty = await Warranty.findByIdAndUpdate(id, {
       ...req.body.warranty,
     });
+    if (!warranty.owner.equals(req.user._doc._id)) {
+      req.flash('error', 'Cannot find the warranty card');
+      return res.redirect('/warrantys');
+    }
     req.flash('success', 'Successfully updated');
     res.redirect('/warrantys');
   })
@@ -90,6 +110,12 @@ router.delete(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const warranty = await Warranty.findById(req.params.id);
+    if (!warranty.owner.equals(req.user._doc._id)) {
+      req.flash('error', 'Cannot find the warranty card');
+      return res.redirect('/warrantys');
+    }
+
     await Warranty.findByIdAndDelete(id);
     res.redirect('/warrantys');
   })
